Poll for memory every 30 seconds instead of every 1/4 second

+++ merb-core/lib/merb-core/rack/adapter/abstract.rb
@@ -152,12 +152,13 @@ module Merb
                 pid, status = @pids[port + i], nil
                 poller = Merb::System::PortablePoller.new(pid)
                 begin
+                  i = 0
                   loop do                    
                     # Watch for the pid to exit.
                     _, status = Process.wait2(pid, Process::WNOHANG)
                     break if status
                     
-                    if Merb::Config[:max_memory] && poller.memory > Merb::Config[:max_memory]
+                    if (i % 120 == 0) && Merb::Config[:max_memory] && poller.memory > Merb::Config[:max_memory]
                       Process.kill("INT", pid)
                       if (Process.kill(0, pid) rescue false)
                         sleep Merb::Config[:hang_time] || 5
@@ -168,6 +169,7 @@ module Merb
                       status = Struct.new(:exitstatus).new(nil)
                       break
                     end
+                    i += 1
                     sleep 0.25
                   end
